+++ src/Controller/VideosController.php
@@ -5,34 +5,83 @@ namespace App\Controller;
 use App\Model\TypesManager;
 use App\Model\VideosManager;
 use App\Model\CategoriesManager;
-use App\Services\FileUploadService;
+use App\Trait\MediasTrait;
 
 class VideosController extends AbstractController
 {
+    use MediasTrait;
+
     /**
      * List Films
      */
-    public function index(): string
+    public function index(?string $category = null, ?string $type = null): string
     {
         $categoriesManager = new CategoriesManager();
         $typeManager = new TypesManager();
         $videosManager = new VideosManager();
-        $medias = $videosManager->selectAll('title');
 
+        $medias = $this->getMedias($videosManager, $category, $type);
+
+        $this->addCategoriesAndTypes($medias, $categoriesManager, $typeManager);
+
+        $title = "Films - Séries - Jeunesses - Documentaires";
+        $categoryFilters = array_merge(['Tout'], $categoriesManager->getAllVideoCategories());
+        $typeFilters = array_merge(['Tout'], $typeManager->getAllTypes());
+
+        return $this->twig->render('Media/index.html.twig', [
+            'page_title' => $title,
+            'categoryFilters' => $categoryFilters,
+            'typeFilters' => $typeFilters,
+            'medias' => $medias,
+            'media_type' => 'videos',
+            'selected_category' => $category,
+            'selected_type' => $type
+        ]);
+    }
+
+    private function getMedias(VideosManager $videosManager, ?string $category, ?string $type): array
+    {
+        if ($this->hasCategoryAndType($category, $type)) {
+            return $videosManager->selectByCategoryAndType($category, $type);
+        }
+
+        if ($this->hasCategory($category)) {
+            $categoryFullName = 'Video ' . $category;
+            return $videosManager->selectByCategory($categoryFullName);
+        }
+
+        if ($this->hasType($type)) {
+            return $videosManager->selectByType($type);
+        }
+
+        return $videosManager->selectAll('title');
+    }
+
+
+    private function addCategoriesAndTypes(
+        array &$medias,
+        CategoriesManager $categoriesManager,
+        TypesManager $typeManager
+    ): void {
         foreach ($medias as &$media) {
             $media['categories'] = $categoriesManager->getCategoriesByVideoId($media['id']);
             $media['types'] = $typeManager->getTypesByVideoId($media['id']);
         }
+    }
 
-        $title = "Films";
-        $filters = ['Action', 'Comédie', 'Drame', 'Documentaire', 'Science-fiction', 'Horreur'];
+    private function hasCategoryAndType(?string $category, ?string $type): bool
+    {
+        return $category && $category !== 'Tout' && $type && $type !== 'Tout';
+    }
 
-        return $this->twig->render('Media/index.html.twig', [
-            'page_title' => $title,
-            'filters' => $filters,
-            'medias' => $medias,
-            'media_type' => 'videos'
-        ]);
+    private function hasCategory(?string $category): bool
+    {
+        return $category && $category !== 'Tout';
+    }
+
+    private function hasType(?string $type): bool
+    {
+        return $type && $type !== 'Tout';
     }
 
     /**
@@ -57,42 +106,44 @@ class VideosController extends AbstractController
         $categories = $categoriesManager->selectAll();
         $typeManager = new TypesManager();
         $types = $typeManager->selectAll();
+        $userRole = $this->getUserRole();
 
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            // clean $_POST data
-            $media = array_map('trim', $_POST);
+        if ($userRole === 'admin') {
+            if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                // clean $_POST data
+                $media = array_map('trim', $_POST);
 
-            $errors = $this->validate($media);
+                $errors = $this->validate($media);
 
-            // Validation de la catégorie
-            if (empty($media['id_types'])) {
-                $errors['id_types'] = 'Le type est requis.';
-            } elseif (!is_numeric($media['id_types'])) {
-                $errors['id_types'] = 'Identifiant de type invalide.';
-            }
-
-            // Si aucune erreur, procéder à l'insertion
-            if (empty($errors)) {
-                $uploadService = new FileUploadService();
-                $fileName = $uploadService->uploadFile($errors);
-                if ($fileName !== "") {
-                    $media['picture'] = $fileName;
-                } else {
-                    $media['picture'] = null;
+                // Validation de la catégorie
+                if (empty($media['id_types'])) {
+                    $errors['id_types'] = 'Le type est requis.';
+                } elseif (!is_numeric($media['id_types'])) {
+                    $errors['id_types'] = 'Identifiant de type invalide.';
                 }
 
-                $videosManager->update($media);
-                header('Location:/videos/show?id=' . $id);
-                return null;
+                // Si aucune erreur, procéder à l'insertion
+                if (empty($errors)) {
+                    $videosManager->update($media);
+                    header('Location:/videos/show?id=' . $id);
+                    return null;
+                }
+
+                // Renvoyer le formulaire avec les erreurs et les données saisies
+                return $this->twig->render('Media/edit.html.twig', [
+                    'categories' => $categories, 'types' => $types,
+                    'errors' => $errors, 'media' => $media, 'media_type' => 'videos', 'isEdit' => true
+                ]);
             }
 
-            // Renvoyer le formulaire avec les erreurs et les données saisies
-            return $this->twig->render('Media/edit.html.twig', ['categories' => $categories, 'types' => $types,
-                'errors' => $errors, 'media' => $media, 'media_type' => 'videos', 'isEdit' => true]);
+            return $this->twig->render('Media/edit.html.twig', [
+                'categories' => $categories, 'types' => $types,
+                'media' => $media, 'media_type' => 'videos', 'isEdit' => true
+            ]);
         }
 
-        return $this->twig->render('Media/edit.html.twig', ['categories' => $categories, 'types' => $types,
-            'media' => $media, 'media_type' => 'videos', 'isEdit' => true]);
+        header('Location:/videos');
+        return null;
     }
 
     /**
@@ -104,43 +155,45 @@ class VideosController extends AbstractController
         $categories = $categoriesManager->selectAll();
         $typeManager = new TypesManager();
         $types = $typeManager->selectAll();
+        $userRole = $this->getUserRole();
 
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            // Nettoyer les données POST
-            $media = array_map('trim', $_POST);
+        if ($userRole === 'admin') {
+            if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                // Nettoyer les données POST
+                $media = array_map('trim', $_POST);
 
-            $errors = $this->validate($media);
+                $errors = $this->validate($media);
 
-            // Validation de la catégorie
-            if (empty($media['id_type'])) {
-                $errors['id_type'] = 'Le type est requis.';
-            } elseif (!is_numeric($media['id_type'])) {
-                $errors['id_type'] = 'Identifiant de type invalide.';
-            }
-
-            // Si aucune erreur, procéder à l'insertion
-            if (empty($errors)) {
-                $uploadService = new FileUploadService();
-                $fileName = $uploadService->uploadFile($errors);
-                if ($fileName !== "") {
-                    $media['picture'] = $fileName;
-                } else {
-                    $media['picture'] = null;
+                // Validation de la catégorie
+                if (empty($media['id_types'])) {
+                    $errors['id_types'] = 'Le type est requis.';
+                } elseif (!is_numeric($media['id_types'])) {
+                    $errors['id_types'] = 'Identifiant de type invalide.';
                 }
 
-                $videosManager = new VideosManager();
-                $id = $videosManager->insert($media);
-                header('Location:/videos/show?id=' . $id);
-                return null;
+                // Si aucune erreur, procéder à l'insertion
+                if (empty($errors)) {
+                    $videosManager = new VideosManager();
+                    $id = $videosManager->insert($media);
+                    header('Location:/videos/show?id=' . $id);
+                    return null;
+                }
+
+                // Renvoyer le formulaire avec les erreurs et les données saisies
+                return $this->twig->render('Media/add.html.twig', [
+                    'categories' => $categories, 'types' => $types,
+                    'errors' => $errors, 'media' => $media, 'media_type' => 'videos'
+                ]);
             }
 
-            // Renvoyer le formulaire avec les erreurs et les données saisies
-            return $this->twig->render('Media/add.html.twig', ['categories' => $categories, 'types' => $types,
-                'errors' => $errors, 'media' => $media, 'media_type' => 'videos']);
+            return $this->twig->render('Media/add.html.twig', [
+                'categories' => $categories, 'types' => $types,
+                'media_type' => 'videos'
+            ]);
         }
 
-        return $this->twig->render('Media/add.html.twig', ['categories' => $categories, 'types' => $types,
-            'media_type' => 'videos']);
+        header('Location:/videos');
+        return null;
     }
 
     /**
@@ -148,18 +201,24 @@ class VideosController extends AbstractController
      */
     public function delete(): void
     {
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $id = trim($_POST['id']);
-            $videosManager = new VideosManager();
-            $video = $videosManager->selectOneById((int)$id);
-            $fileName = "../public/assets/images/covers/" . $video['picture'];
+        $userRole = $this->getUserRole();
 
-            if ($video['picture'] && file_exists($fileName)) {
-                unlink($fileName);  // Supprime le fichier image
+        if ($userRole === 'admin') {
+            if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                $id = trim($_POST['id']);
+                $videosManager = new VideosManager();
+                $video = $videosManager->selectOneById((int)$id);
+                $fileName = "../public/assets/images/covers/" . $video['picture'];
+
+                if ($video['picture'] && file_exists($fileName)) {
+                    unlink($fileName);  // Supprime le fichier image
+                }
+
+                $videosManager->delete((int)$id);
+                header('Location:/videos');
             }
-
-            $videosManager->delete((int)$id);
-            header('Location:/books');
         }
+
+        header('Location:/videos');
     }
 }
